Move generateCertificateInterest and encryptAndPublish
