+++ app/src/main/java/memphis/myapplication/Common.java
@@ -1,22 +1,61 @@
 package memphis.myapplication;
 
+import android.content.Context;
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.net.Uri;
+import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.intel.jndn.management.ManagementException;
+import com.intel.jndn.management.Nfdc;
+
+import io.realm.Realm;
+import memphis.myapplication.RealmObjects.PublishedContent;
+import memphis.myapplication.RealmObjects.SelfCertificate;
+import memphis.myapplication.RealmObjects.User;
 import timber.log.Timber;
 
 import net.named_data.jndn.ContentType;
 import net.named_data.jndn.Data;
+import net.named_data.jndn.Interest;
 import net.named_data.jndn.MetaInfo;
 import net.named_data.jndn.Name;
+import net.named_data.jndn.OnData;
+import net.named_data.jndn.OnTimeout;
 import net.named_data.jndn.encoding.EncodingException;
 
+import net.named_data.jndn.encrypt.algo.EncryptAlgorithmType;
+import net.named_data.jndn.encrypt.algo.EncryptParams;
+import net.named_data.jndn.encrypt.algo.RsaAlgorithm;
 import net.named_data.jndn.security.KeyChain;
 import net.named_data.jndn.security.SecurityException;
+import net.named_data.jndn.security.VerificationHelpers;
 import net.named_data.jndn.security.pib.PibImpl;
 import net.named_data.jndn.security.tpm.TpmBackEnd;
+import net.named_data.jndn.security.v2.CertificateV2;
 import net.named_data.jndn.util.Blob;
 
+import org.apache.commons.io.IOUtils;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
 import java.nio.ByteBuffer;
 
+import java.security.InvalidAlgorithmParameterException;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+
+import javax.crypto.BadPaddingException;
+import javax.crypto.IllegalBlockSizeException;
+import javax.crypto.NoSuchPaddingException;
+import javax.crypto.SecretKey;
+
+import static memphis.myapplication.Globals.consumerManager;
+import static memphis.myapplication.Globals.producerManager;
 
 
 public class Common {
@@ -107,6 +146,253 @@ public class Common {
             segment_number++;
         } while (byteBuffer.hasRemaining());
         return datas;
+    }
+
+    /**
+     * Generates and expresses interest for our certificate signed by friend
+     * @param friend: name of friend who has our certificate
+     */
+    static void generateCertificateInterest(String friend) throws SecurityException, IOException {
+        Realm realm = Realm.getDefaultInstance();
+        User user = realm.where(User.class).equalTo("username", friend).findFirst();
+        Name name =  new Name(user.getNamespace());
+        name.append("cert");
+        Name certName = Globals.keyChain.getDefaultCertificateName();
+        Name newCertName = new Name();
+        int end = 0;
+        for (int i = 0; i<= certName.size(); i++) {
+            if (certName.getSubName(i, 1).toUri().equals("/self")) {
+                newCertName.append(certName.getPrefix(i));
+                end = i;
+                break;
+            }
+        }
+        newCertName.append(friend);
+        newCertName.append(certName.getSubName(end+1));
+        name.append(newCertName);
+        Interest interest = new Interest(name);
+        Timber.d("Expressing interest for our cert %s", name.toUri());
+        registerUser(friend);
+        Globals.face.expressInterest(interest, onCertData, onCertTimeOut);
+    }
+
+    public static void registerUser(String friendName) {
+        Realm realm = Realm.getDefaultInstance();
+        User friend = realm.where(User.class).equalTo("username", friendName).findFirst();
+        try {
+            Nfdc.register(Globals.face, Globals.multicastFaceID, new Name(friend.getNamespace()), 0);
+        } catch (ManagementException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Callback for certificate from friend
+     */
+    static OnData onCertData = new OnData() {
+
+        @Override
+        public void onData(Interest interest, Data data) {
+            Timber.d("Getting our certificate back from friend");
+            Realm realm = Realm.getDefaultInstance();
+
+            String friendName = interest.getName().getSubName(-2, 1).toUri().substring(1);
+            User friend = realm.where(User.class).equalTo("username", friendName).findFirst();
+            Blob interestData = data.getContent();
+            byte[] certBytes = interestData.getImmutableArray();
+
+            CertificateV2 certificateV2 = new CertificateV2();
+            try {
+                certificateV2.wireDecode(ByteBuffer.wrap(certBytes));
+            } catch (EncodingException e) {
+                e.printStackTrace();
+            }
+            realm.beginTransaction();
+            SelfCertificate realmCertificate = realm.where(SelfCertificate.class).equalTo("username", friendName).findFirst();
+            if (realmCertificate == null) {
+                realmCertificate = realm.createObject(SelfCertificate.class, friendName);
+            }
+            realmCertificate.setCert(certificateV2);
+
+            VerificationHelpers verificationHelpers = new VerificationHelpers();
+            try {
+                boolean verified = verificationHelpers.verifyDataSignature(certificateV2, realm.where(User.class).equalTo("username", friendName).findFirst().getCert());
+            } catch (EncodingException e) {
+                e.printStackTrace();
+            }
+
+            Timber.d("Saved our certificate back signed by friend and adding them as a consumer");
+
+            friend.setFriend(true);
+            friend.setTrust(true);
+            consumerManager.createConsumer(friend.getNamespace());
+            realm.commitTransaction();
+
+            // Share friend's list
+            producerManager.updateFriendsList();
+
+            if (!Globals.useMulticast) {
+                Globals.nsdHelper.registerUser(friendName);
+            } else {
+                User user = realm.where(User.class).equalTo("username", friendName).findFirst();
+                try {
+                    Nfdc.register(Globals.face, Globals.multicastFaceID, new Name(user.getNamespace()), 0);
+                } catch (ManagementException e) {
+                    e.printStackTrace();
+                }
+
+            }
+            realm.close();
+        }
+    };
+
+    /**
+     * Callback for timeout of interest for certificate from friend
+     */
+    static OnTimeout onCertTimeOut = new OnTimeout() {
+
+        @Override
+        public void onTimeout(Interest interest) {
+            Timber.d( "Timeout for interest " + interest.toUri());
+            String friend = interest.getName().getSubName(-2, 1).toString().substring(1);
+            Timber.d("Resending interest to " + friend);
+            try {
+                generateCertificateInterest(friend);
+            } catch (SecurityException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    };
+
+
+    /**
+     * encodes sync data, encrypts photo, and publishes filename and symmetric keys
+     * @param resultData: intent with filename and recipients list
+     */
+    public static void encryptAndPublish(Intent resultData, Context context) {
+        try {
+            final String path = resultData.getStringExtra("photo");
+            final File photo = new File(path);
+            Timber.d("File size: " + photo.length());
+            final Uri uri = UriFileProvider.getUriForFile(context,
+                    context.getApplicationContext().getPackageName() +
+                            ".UriFileProvider", photo);
+            final Encrypter encrypter = new Encrypter(context.getApplicationContext());
+
+
+            ArrayList<String> recipients;
+            try {
+                recipients = resultData.getStringArrayListExtra("recipients");
+                SharedPrefsManager sharedPrefsManager = SharedPrefsManager.getInstance(context);
+
+                String name = sharedPrefsManager.getNamespace() + "/data";
+                final String filename = sharedPrefsManager.getNamespace() + "/file" + path;
+
+                // Generate symmetric key
+                final SecretKey secretKey = encrypter.generateKey();
+                final byte[] iv = encrypter.generateIV();
+
+                // Encode sync data
+                SyncData syncData = new SyncData();
+                syncData.setFilename(filename);
+
+                final boolean feed = (recipients == null);
+                if (feed) {
+                    Timber.d("For feed");
+                    syncData.setFeed(true);
+                }
+                else {
+                    syncData.setFeed(false);
+                    Timber.d( "For friends");
+                    Realm realm = Realm.getDefaultInstance();
+                    for (String friend : recipients) {
+                        Blob friendKey = realm.where(User.class).equalTo("username", friend).findFirst().getCert().getPublicKey();
+                        byte[] encryptedKey = RsaAlgorithm.encrypt
+                                (friendKey, new Blob(secretKey.getEncoded()), new EncryptParams(EncryptAlgorithmType.RsaOaep)).getImmutableArray();
+                        syncData.addFriendKey(friend, encryptedKey);
+                    }
+                }
+                // Stringify sync data
+                producerManager.setDataSeqMap(syncData.stringify());
+                Timber.d("Publishing file: %s", filename);
+
+                byte[] bytes;
+                try {
+                    InputStream is = context.getContentResolver().openInputStream(uri);
+                    bytes = IOUtils.toByteArray(is);
+                    Timber.d("select file activity: %s", "file byte array size: " + bytes.length);
+                } catch (IOException e) {
+                    Timber.d("onItemClick: failed to byte");
+                    e.printStackTrace();
+                    bytes = new byte[0];
+                }
+                Timber.d("file selection result: %s", "file path: " + path);
+                try {
+                    String prefixApp = "/" + sharedPrefsManager.getNamespace();
+
+                    final String prefix = prefixApp + "/file" + path;
+                    Timber.d(prefix);
+                    Realm realm = Realm.getDefaultInstance();
+                    realm.beginTransaction();
+                    PublishedContent contentKey = realm.createObject(PublishedContent.class, path);
+                    if (!feed) {
+                        Timber.d("Publishing to friend(s)");
+                        contentKey.addKey(secretKey);
+                        realm.commitTransaction();
+                        realm.close();
+
+                        Blob encryptedBlob = encrypter.encrypt(secretKey, iv, bytes);
+                        Common.publishData(encryptedBlob, new Name(prefix));
+                    }
+                    else {
+                        Timber.d("Publishing to feed");
+                        realm.commitTransaction();
+                        realm.close();
+                        Blob unencryptedBlob = new Blob(bytes);
+                        Common.publishData(unencryptedBlob, new Name(prefix));
+
+                    }
+                    final FileManager manager = new FileManager(context.getApplicationContext());
+                    Bitmap bitmap = QRExchange.makeQRCode(prefix);
+                    manager.saveFileQR(bitmap, prefix);
+                    ((AppCompatActivity)context).runOnUiThread(makeToast("Sending photo", context));
+                } catch (NoSuchAlgorithmException e) {
+                    e.printStackTrace();
+                } catch (NoSuchPaddingException e) {
+                    e.printStackTrace();
+                } catch (BadPaddingException e) {
+                    e.printStackTrace();
+                } catch (InvalidKeyException e) {
+                    e.printStackTrace();
+                } catch (IllegalBlockSizeException e) {
+                    e.printStackTrace();
+                } catch (InvalidAlgorithmParameterException e) {
+                    e.printStackTrace();
+                }
+                producerManager.publishFile(name);
+            }
+            catch(Exception e) {
+                e.printStackTrace();
+            }
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+            ((AppCompatActivity)context).runOnUiThread(makeToast("Something went wrong with sending photo. Try resending", context));
+        }
+    }
+
+    /**
+     * Android is very particular about UI processes running on a separate thread. This function
+     * creates and returns a Runnable thread object that will display a Toast message.
+     */
+    public static Runnable makeToast(final String s, final Context context) {
+        return new Runnable() {
+            public void run() {
+                Toast.makeText(context, s, Toast.LENGTH_LONG).show();
+            }
+        };
     }
 
 }
