+++ app/src/main/java/memphis/myapplication/Globals.java
@@ -82,125 +82,6 @@ public class Globals extends Application {
         }
     }
 
-    /**
-     * Generates and expresses interest for our certificate signed by friend
-     * @param friend: name of friend who has our certificate
-     */
-    static void generateCertificateInterest(String friend) throws SecurityException, IOException {
-        Realm realm = Realm.getDefaultInstance();
-        User user = realm.where(User.class).equalTo("username", friend).findFirst();
-        Name name =  new Name(user.getNamespace());
-        name.append("cert");
-        Name certName = keyChain.getDefaultCertificateName();
-        Name newCertName = new Name();
-        int end = 0;
-        for (int i = 0; i<= certName.size(); i++) {
-            if (certName.getSubName(i, 1).toUri().equals("/self")) {
-                newCertName.append(certName.getPrefix(i));
-                end = i;
-                break;
-            }
-        }
-        newCertName.append(friend);
-        newCertName.append(certName.getSubName(end+1));
-        name.append(newCertName);
-        Interest interest = new Interest(name);
-        Timber.d("Expressing interest for our cert %s", name.toUri());
-        registerUser(friend);
-        face.expressInterest(interest, onCertData, onCertTimeOut);
-    }
-
-    public static void registerUser(String friendName) {
-        Realm realm = Realm.getDefaultInstance();
-        User friend = realm.where(User.class).equalTo("username", friendName).findFirst();
-        try {
-            Nfdc.register(face, Globals.multicastFaceID, new Name(friend.getNamespace()), 0);
-        } catch (ManagementException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Callback for certificate from friend
-     */
-    static OnData onCertData = new OnData() {
-
-        @Override
-        public void onData(Interest interest, Data data) {
-            Timber.d("Getting our certificate back from friend");
-            Realm realm = Realm.getDefaultInstance();
-
-            String friendName = interest.getName().getSubName(-2, 1).toUri().substring(1);
-            User friend = realm.where(User.class).equalTo("username", friendName).findFirst();
-            Blob interestData = data.getContent();
-            byte[] certBytes = interestData.getImmutableArray();
-
-            CertificateV2 certificateV2 = new CertificateV2();
-            try {
-                certificateV2.wireDecode(ByteBuffer.wrap(certBytes));
-            } catch (EncodingException e) {
-                e.printStackTrace();
-            }
-            realm.beginTransaction();
-            SelfCertificate realmCertificate = realm.where(SelfCertificate.class).equalTo("username", friendName).findFirst();
-            if (realmCertificate == null) {
-                realmCertificate = realm.createObject(SelfCertificate.class, friendName);
-            }
-            realmCertificate.setCert(certificateV2);
-
-            VerificationHelpers verificationHelpers = new VerificationHelpers();
-            try {
-                boolean verified = verificationHelpers.verifyDataSignature(certificateV2, realm.where(User.class).equalTo("username", friendName).findFirst().getCert());
-            } catch (EncodingException e) {
-                e.printStackTrace();
-            }
-
-            Timber.d("Saved our certificate back signed by friend and adding them as a consumer");
-
-            friend.setFriend(true);
-            friend.setTrust(true);
-            consumerManager.createConsumer(friend.getNamespace());
-            realm.commitTransaction();
-
-            // Share friend's list
-            producerManager.updateFriendsList();
-
-            if (!Globals.useMulticast) {
-                Globals.nsdHelper.registerUser(friendName);
-            } else {
-                User user = realm.where(User.class).equalTo("username", friendName).findFirst();
-                try {
-                    Nfdc.register(face, Globals.multicastFaceID, new Name(user.getNamespace()), 0);
-                } catch (ManagementException e) {
-                    e.printStackTrace();
-                }
-
-            }
-            realm.close();
-        }
-    };
-
-    /**
-     * Callback for timeout of interest for certificate from friend
-     */
-    static OnTimeout onCertTimeOut = new OnTimeout() {
-
-        @Override
-        public void onTimeout(Interest interest) {
-            Timber.d( "Timeout for interest " + interest.toUri());
-            String friend = interest.getName().getSubName(-2, 1).toString().substring(1);
-            Timber.d("Resending interest to " + friend);
-            try {
-                generateCertificateInterest(friend);
-            } catch (SecurityException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-    };
-
-
     //v2 changes
     public static void setPib(AndroidSqlite3Pib p) {
         pib = p;
